Fix Windows build error and MSVC warnings
Add missing <stdexcept> include in UserPort.cpp which MSVC requires
(GCC/Clang pull it in transitively). Add static_casts to silence C4244
narrowing warnings: int-to-char in std::tolower lambdas, uint64-to-uint32
in Expression::emit_const call.

+++ src/core/src/Expression.cpp
@@ -311,7 +311,7 @@ struct Parser {
         }
 
         (void)start;
-        emit_const(value);
+        emit_const(static_cast<uint32_t>(value));
     }
 };
